Add files via upload
There are 2 inputs & 1 output in this model. Each input use trimff methods for fuzzification, and  have 3 membership function(N-negative, Z-zero, P-positive ),the defuzzification method is weight average method.
This is model is used for further reinforcement learning, and the simulation environment should be settled in simulink. That is why i can not using the fuzzy tool box in matlab（the fuzzy parameters can not update real-time）.
I am really a green hand in this domine, any suggestion for Fuzzy control or RL will be welcome~